functions/requirements.txt
firebase-functions
google-cloud-firestore
firebase-admin